# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: Creation Agent Studio API
  version: 2.0.0
  description: |
    Creation Agent Studio backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/v2/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List openable applications (bare array, no pagination envelope)
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [chat, task, custom, all], default: chat }
          description: Invalid values fall back to chat.
        - name: scope
          in: query
          schema: { type: string, enum: [public, mine, manage], default: public }
          description: Invalid values fall back to public.
        - name: include_unbound
          in: query
          schema: { type: string }
          description: Truthy when case-insensitively in {1,true,yes}.
      responses:
        "200":
          description: Application list items (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApplicationListItem" }
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create an application (admin only — agent marketplace "new Aily agent")
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationCreateRequest" }
      responses:
        "201":
          description: Created (compact detail shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
        "403":
          description: only admins may add agents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      operationId: getApplication
      tags: [applications]
      summary: Application detail (compact shape)
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application / rebind runtime (creator or staff)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationUpdateRequest" }
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "403":
          description: only creator or staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application (409 when referenced by conversations/runs)
      responses:
        "204":
          description: Deleted
        "403":
          description: only creator or staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "409":
          description: has conversations or runs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}/favorite:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: favoriteApplication
      tags: [applications]
      summary: Pin application (idempotent)
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteResult" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: unfavoriteApplication
      tags: [applications]
      summary: Unpin application (idempotent)
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteResult" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}/avatar:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      operationId: getApplicationAvatar
      tags: [applications]
      summary: Stream avatar image bytes
      responses:
        "200":
          description: Image bytes
          content:
            image/png: {}
            image/jpeg: {}
            image/gif: {}
            image/webp: {}
        "404":
          description: avatar not set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    post:
      operationId: uploadApplicationAvatar
      tags: [applications]
      summary: Replace avatar (multipart file; png/jpg/jpeg/gif/webp ≤ 2MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
    delete:
      operationId: clearApplicationAvatar
      tags: [applications]
      summary: Clear avatar (falls back to emoji icon)
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
  /api/v2/applications/{id}/default-agent:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: setDefaultAgent
      tags: [applications]
      summary: Set as workspace main agent (chat + enabled binding only)
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: not eligible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "403":
          description: only creator or staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: unsetDefaultAgent
      tags: [applications]
      summary: Unset workspace main agent
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
...
components:
...
  schemas:
...
    DetailError:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
...
    ApplicationListItem:
      type: object
      description: shape returned by GET /api/v2/applications
      required:
        - id
        - slug
        - name
        - description
        - icon
        - avatar_url
        - color
        - kind
        - renderer_key
        - executor_key
        - category_slug
        - category_name
        - is_public
        - enabled
        - runtime_type
        - provider_key
        - external_resource_id
        - identity_mode
        - execution_mode
        - capabilities
        - is_bound
        - is_favorite
        - is_default_agent
        - can_manage
        - usage_count
        - last_used_at
        - global_usage_count
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        avatar_url: { type: string }
        color: { type: string }
        kind: { type: string, enum: [chat, task, custom] }
        renderer_key: { type: string }
        executor_key: { type: string }
        category_slug: { type: string }
        category_name: { type: string }
        is_public: { type: boolean }
        enabled: { type: boolean }
        runtime_type: { type: string }
        provider_key: { type: string }
        external_resource_id: { type: string }
        identity_mode: { type: string }
        execution_mode: { type: string }
        capabilities:
          type: object
          additionalProperties: { type: boolean }
        is_bound: { type: boolean }
        is_favorite: { type: boolean }
        is_default_agent: { type: boolean }
        can_manage: { type: boolean }
        usage_count: { type: integer }
        last_used_at:
          type: string
          nullable: true
          format: date-time
        global_usage_count: { type: integer }
    ApplicationDetail:
      type: object
      description: compact shape returned by authoring / avatar / default-agent endpoints
      required:
        - id
        - slug
        - name
        - description
        - icon
        - avatar_url
        - color
        - kind
        - is_public
        - enabled
        - category_slug
        - category_name
        - is_default_agent
        - is_bound
        - runtime_type
        - provider_key
        - external_resource_id
        - identity_mode
        - execution_mode
        - can_manage
        - updated_at
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        avatar_url: { type: string }
        color: { type: string }
        kind: { type: string, enum: [chat, task, custom] }
        is_public: { type: boolean }
        enabled: { type: boolean }
        category_slug: { type: string }
        category_name: { type: string }
        is_default_agent: { type: boolean }
        is_bound: { type: boolean }
        runtime_type: { type: string }
        provider_key: { type: string }
        external_resource_id: { type: string }
        identity_mode: { type: string }
        execution_mode: { type: string }
        can_manage: { type: boolean }
        updated_at: { type: string, format: date-time }
...
    ApplicationCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100, minLength: 1 }
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description: { type: string }
        icon: { type: string, maxLength: 50 }
        color: { type: string, maxLength: 20 }
        category_slug: { type: string, maxLength: 100 }
        category_name: { type: string, maxLength: 100 }
        is_public: { type: boolean, default: true }
        kind: { type: string, enum: [chat], default: chat }
        renderer_key: { type: string, default: chat }
        runtime: { $ref: "#/components/schemas/RuntimeBindingInput" }
        set_default_agent: { type: boolean, default: false }
...